+++ clients/derivatives_trading_usds_futures/src/binance_sdk_derivatives_trading_usds_futures/websocket_api/websocket_api.py
@@ -449,7 +449,7 @@ class DerivativesTradingUsdsFuturesWebSocketAPI(WebSocketAPIBase):
         symbol: Union[str, None],
         side: Union[ModifyOrderSideEnum, None],
         quantity: Union[float, None],
-        price: Union[float, None],
+        price: Optional[float] = None,
         id: Optional[str] = None,
         order_id: Optional[int] = None,
         orig_client_order_id: Optional[str] = None,
@@ -477,7 +477,7 @@ class DerivativesTradingUsdsFuturesWebSocketAPI(WebSocketAPIBase):
                     symbol (Union[str, None]):
                     side (Union[ModifyOrderSideEnum, None]): `SELL`, `BUY`
                     quantity (Union[float, None]): Order quantity, cannot be sent with `closePosition=true`
-                    price (Union[float, None]):
+                    price (Optional[float] = None):
                     id (Optional[str] = None): Unique WebSocket request ID.
                     order_id (Optional[int] = None):
                     orig_client_order_id (Optional[str] = None):